refactor: 예약 대기 서비스 패키지 분리
