+++ src/main/java/roomescape/service/waiting/WaitingService.java
@@ -1,4 +1,4 @@
-package roomescape.service.reservation;
+package roomescape.service.waiting;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
