+++ src/test/java/roomescape/service/payment/PaymentServiceTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import roomescape.domain.payment.Payment;
 import roomescape.domain.dto.PaymentRequest;
+import roomescape.domain.payment.Payment;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
 class PaymentServiceTest {
